refactor : 동일 값 동일 객체 반환
- 동일 값 동일 객체 반환
