+++ src/main/java/lotto/domain/Lotto.java
@@ -7,15 +7,12 @@ import java.util.Set;
 import java.util.TreeSet;
 import java.util.stream.Collectors;
 
-import lotto.domain.number.LottoNumber;
-import lotto.domain.number.LottoNumbers;
 import lotto.enums.Rank;
 import lotto.exception.LottoValidationException;
 
 public class Lotto {
 
 	private static final int LOTTO_NUMBERS_LENGTH = 6;
-	private static final LottoNumbers LOTTO_NUMBERS = LottoNumbers.generate();
 
 	private final Set<LottoNumber> lottoNumbers;
 
@@ -33,7 +30,7 @@ public class Lotto {
 		Set<LottoNumber> lottoNumberList = Arrays.stream(stringLottoNumbers)
 											.map(String::trim)
 											.mapToInt(Integer::parseInt)
-											.mapToObj(LOTTO_NUMBERS::valueOf)
+											.mapToObj(LottoNumber::valueOf)
 											.collect(Collectors.toCollection(TreeSet<LottoNumber>::new));
 
 		return new Lotto(lottoNumberList);
@@ -41,7 +38,7 @@ public class Lotto {
 
 	public static Lotto of(List<Integer> lottoNumbers) {
 		Set<LottoNumber> lottoNumberList = lottoNumbers.stream()
-													.map(LOTTO_NUMBERS::valueOf)
+													.map(LottoNumber::valueOf)
 													.collect(Collectors.toCollection(TreeSet<LottoNumber>::new));
 
 		return new Lotto(lottoNumberList);
